refactor: lift the inline-SVG chart primitives to components/charts.tsx (#428)
The gateway's Usage dashboard (#428) draws the same charts over an entirely
separate table, and charts.tsx is pure presentation over {label, value, hint}[]
with nothing Claude-specific in it. The locked decision that gateway traffic is
never mixed into Claude analytics is about *data* and *sections*; a second copy
of 250 lines of SVG would only guarantee the two drift.

A pure move, on its own commit so the review can see it is one:

- git mv src/views/analytics/charts.tsx -> src/components/charts.tsx.
- CardEmpty moves with it (charts.tsx was its only cross-file consumer besides
  shared.tsx itself) and is re-exported from analytics/shared.tsx, so no
  analytics call site churns.
- The .a-chart*/.a-heat* rules move verbatim out of styles/analytics.css into
  styles/charts.css, imported by the component: a consumer outside
  views/analytics/ gets the styles from the component rather than by importing
  another section's stylesheet. Class names are unchanged, and every selector in
  the new file is unique to it, so its position in the bundle cannot move a
  cascade.

Verified rather than argued: building origin/main and this commit and comparing
the emitted CSS as a sorted set of rules gives 578 identical rules on both
sides, byte for byte.
